requirements.txt
django 
torch 
torchvision 
pillow